+++ tavla/app/(admin)/components/CreateBoard/NameAndFolderSelector.tsx
@@ -5,7 +5,7 @@ import { Heading2, Label, Paragraph } from '@entur/typography'
 import { HiddenInput } from 'app/(admin)/components/Form/HiddenInput'
 import { SubmitButton } from 'app/(admin)/components/Form/SubmitButton'
 import { useFolderDropdown } from 'app/(admin)/hooks/useFolders'
-import { getFormFeedbackForField } from 'app/(admin)/utils'
+import { getFormFeedbackForField } from 'app/(admin)/utils/forms'
 import type { Folder } from 'app/(admin)/utils/types'
 import ClientOnlyTextField from 'app/components/NoSSR/TextField'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
