+++ tavla/app/(admin)/components/CreateFolder/index.tsx
@@ -5,7 +5,7 @@ import { Modal } from '@entur/modal'
 import { Heading2, Label, Paragraph } from '@entur/typography'
 import { SubmitButton } from 'app/(admin)/components/Form/SubmitButton'
 import { FormError } from 'app/(admin)/components/FormError'
-import { getFormFeedbackForField } from 'app/(admin)/utils'
+import { getFormFeedbackForField } from 'app/(admin)/utils/forms'
 import ClientOnlyTextField from 'app/components/NoSSR/TextField'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
 import birds from 'assets/illustrations/Birds.png'
