+++ tavla/app/(admin)/components/Delete/actions.ts
@@ -1,7 +1,10 @@
 'use server'
 
-import { getFormFeedbackForError, type TFormFeedback } from 'app/(admin)/utils'
 import { deleteFolder } from 'app/(admin)/utils/firebase'
+import {
+    getFormFeedbackForError,
+    type TFormFeedback,
+} from 'app/(admin)/utils/forms'
 import { handleError } from 'app/(admin)/utils/handleError'
 import { getUserFromSessionCookie } from 'app/(admin)/utils/server'
 import { revalidatePath } from 'next/cache'
