+++ tavla/app/(admin)/mapper/components/MemberAdministration/InviteUser.tsx
@@ -5,7 +5,7 @@ import { Heading3 } from '@entur/typography'
 import { HiddenInput } from 'app/(admin)/components/Form/HiddenInput'
 import { SubmitButton } from 'app/(admin)/components/Form/SubmitButton'
 import { FormError } from 'app/(admin)/components/FormError'
-import { getFormFeedbackForField } from 'app/(admin)/utils'
+import { getFormFeedbackForField } from 'app/(admin)/utils/forms'
 import ClientOnlyTextField from 'app/components/NoSSR/TextField'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
 import { useActionState, useRef } from 'react'
