+++ tavla/app/(admin)/mapper/components/UploadLogo/DeleteLogo.tsx
@@ -3,7 +3,10 @@ import { useToast } from '@entur/alert'
 import { Button } from '@entur/button'
 import { DeleteIcon } from '@entur/icons'
 import { FormError } from 'app/(admin)/components/FormError'
-import { getFormFeedbackForField, type TFormFeedback } from 'app/(admin)/utils'
+import {
+    getFormFeedbackForField,
+    type TFormFeedback,
+} from 'app/(admin)/utils/forms'
 import { useState } from 'react'
 import type { FolderDB } from 'src/types/db-types/folders'
 import { remove } from './actions'
