+++ tavla/app/(admin)/mapper/components/UploadLogo/actions.ts
@@ -1,12 +1,12 @@
 'use server'
 
 import * as Sentry from '@sentry/nextjs'
-import { getFormFeedbackForError } from 'app/(admin)/utils'
 import {
     getConfig,
     initializeAdminApp,
     userCanEditFolder,
 } from 'app/(admin)/utils/firebase'
+import { getFormFeedbackForError } from 'app/(admin)/utils/forms'
 import { handleError } from 'app/(admin)/utils/handleError'
 import { firestore, storage } from 'firebase-admin'
 import { revalidatePath } from 'next/cache'
