+++ tavla/app/(admin)/oversikt/components/Column/Delete.tsx
@@ -10,7 +10,7 @@ import { HiddenInput } from 'app/(admin)/components/Form/HiddenInput'
 import { SubmitButton } from 'app/(admin)/components/Form/SubmitButton'
 import { FormError } from 'app/(admin)/components/FormError'
 import { deleteBoardAction } from 'app/(admin)/oversikt/utils/actions'
-import { getFormFeedbackForField } from 'app/(admin)/utils'
+import { getFormFeedbackForField } from 'app/(admin)/utils/forms'
 import type { EventProps } from 'app/posthog/events'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
 import sheep from 'assets/illustrations/Sheep.png'
