+++ tavla/app/(admin)/oversikt/utils/actions.ts
@@ -1,8 +1,8 @@
 'use server'
 import { getBoardsForFolder } from 'app/(admin)/actions'
 import { moveBoard } from 'app/(admin)/tavler/[id]/rediger/components/Settings/actions'
-import type { TFormFeedback } from 'app/(admin)/utils'
 import { deleteBoard, initializeAdminApp } from 'app/(admin)/utils/firebase'
+import type { TFormFeedback } from 'app/(admin)/utils/forms'
 import { handleError } from 'app/(admin)/utils/handleError'
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
