+++ tavla/app/(admin)/tavler/[id]/rediger/components/Settings/actions.ts
@@ -6,15 +6,15 @@ import {
     isOnlyWhiteSpace,
 } from 'app/(admin)/tavler/[id]/utils'
 import {
-    getFormFeedbackForError,
-    type InputType,
-    type TFormFeedback,
-} from 'app/(admin)/utils'
-import {
     initializeAdminApp,
     userCanEditBoard,
     userCanEditFolder,
 } from 'app/(admin)/utils/firebase'
+import {
+    getFormFeedbackForError,
+    type InputType,
+    type TFormFeedback,
+} from 'app/(admin)/utils/forms'
 import { handleError } from 'app/(admin)/utils/handleError'
 import { getUserFromSessionCookie } from 'app/(admin)/utils/server'
 import { FieldValue, getFirestore } from 'firebase-admin/firestore'
