+++ tavla/app/(admin)/tavler/[id]/rediger/components/TileCard/components/SaveCancelDeleteTileButtonGroup.tsx
@@ -5,7 +5,7 @@ import { Modal } from '@entur/modal'
 import { Heading3, Paragraph } from '@entur/typography'
 import { SubmitButton } from 'app/(admin)/components/Form/SubmitButton'
 import { TileContext } from 'app/(admin)/tavler/[id]/rediger/components/TileCard/context'
-import type { TFormFeedback } from 'app/(admin)/utils'
+import type { TFormFeedback } from 'app/(admin)/utils/forms'
 import type { EventProps } from 'app/posthog/events'
 import { usePosthogTracking } from 'app/posthog/usePosthogTracking'
 import Goat from 'assets/illustrations/Goat.png'
