+++ tavla/app/(admin)/utils/handleError.ts
@@ -1,6 +1,6 @@
 import { FirebaseError } from 'firebase/app'
 import { isString } from 'lodash'
-import { getFormFeedbackForError } from '.'
+import { getFormFeedbackForError } from './forms'
 
 export function handleError(e: unknown) {
     if (e instanceof FirebaseError || isString(e)) {
